// src/lib/api-client/types.ts
export interface components {
    schemas: {
// ...
        Wallet: {
            /** Format: uuid */
            id?: string;
            /** Format: uuid */
            userId?: string;
            balance?: {
                /** @example 1000 */
                amount?: number;
                /** @example BRL */
                currency?: string;
            };
            /** Format: date-time */
            createdAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}